requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
psutil==5.9.8
requests
